Fix broken tests (#1178)
* remove hardcoded attributes override


* handle failure without rootCause

throw an java.lang.AssertionError in case the expected and actual response codes do not match.


* handle index out of bounds exception


* hashing lambda tests


* increasing timeout for native android tests

this is needed to compensate the lack of caching


* adding retry logic for browser creation

- will retry only once and with a static polling interval of 5 seconds (5000 millis).
- TODO: make retry attempts and polling interval configurable, and globalize it for all driver instances (it's currently only for local driver creation)


* separating API tests to an independent job


* amending scope for API test job


* fix scope to skip api tests for invalids

- Windows_Chrome_Local
- Windows_Edge_Local


* convert timeout property from int to double

- this allows us to configure the timeout to 0.6 in case of mobile scrolling.
- this fixes an issue that has laid dormant since the built-in properties' migration.


* fix pipeline syntax issue


---------

+++ src/main/java/com/shaft/gui/element/AlertActions.java
@@ -41,7 +41,7 @@ public class AlertActions {
 
     private static void waitForAlertToBePresent() {
         try {
-            (new WebDriverWait(DriverFactoryHelper.getDriver().get(), Duration.ofSeconds(SHAFT.Properties.timeouts.defaultElementIdentificationTimeout()))).until(ExpectedConditions.alertIsPresent());
+            (new WebDriverWait(DriverFactoryHelper.getDriver().get(), Duration.ofSeconds((long) SHAFT.Properties.timeouts.defaultElementIdentificationTimeout()))).until(ExpectedConditions.alertIsPresent());
             DriverFactoryHelper.getDriver().get().switchTo().alert();
             ReportManager.logDiscrete("Alert is present");
         } catch (Exception rootCauseException) {
